Draft: Tc2 8 alt

See merge request foxminded-teamconnect-2/tracker!9

+++ src/main/java/com/ua/teamconnect/tracker/model/dto/DepartmentDto.java
@@ -5,10 +5,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(name = "Department", description = "Data of the department")
 public record DepartmentDto(
     @Schema(description = "Id of the department", example = "1")
-    Long id,
+    Integer id,
     @Schema(description = "Name of the department", example = "Software Development")
     String name,
     @Schema(description = "Head Id of this department", example = "1")
-    Long headId
+    Integer headId
 ) {
 }
