+++ src/main/java/com/ua/teamconnect/tracker/repository/StackRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StackRepository extends JpaRepository<Stack, Long> {
+public interface StackRepository extends JpaRepository<Stack, Integer> {
 }
